observed blue/green failover did not work
This change accounts for blue/green failover not working due to using a
connection pool with persistent connections to call out to the service.

https://go-review.googlesource.com/c/go/+/22670
There is a built in IdleConnTimeout on the transport, but it doesn't
solve the problem observed.

Since we're using keep-alives and we have a high request rate we never
leave the connection idle long enough to be recycled. A fix for this is
to force connection resets on a cycle.

I would be interested in trying some other ingress solutions to see how
/ if they handle this.